Removed unused variables and code optimization.
