+++ src/parser.rs
@@ -1,89 +1,6 @@
+use crate::types::{ArgTuple, Argument, Callback};
 use log::{debug, error};
-use rocket::tokio;
-use std::{
-    future::Future,
-    io::{Error, ErrorKind},
-    pin::Pin,
-};
-
-use crate::{types::Message, WebexClient};
-
-// ###########################################################################
-// Tuple definition that contains the name:value mapping.
-// ###########################################################################
-
-pub type ArgTuple = Vec<(std::string::String, std::string::String)>;
-pub type Callback = fn(
-    WebexClient,
-    Message,
-    ArgTuple,
-    ArgTuple,
-) -> (Pin<Box<dyn Future<Output = ()> + Send + Sync + 'static>>);
-
-// ###################################################################
-// Define the Argument trait
-// ###################################################################
-
-pub trait Argument: Send + Sync {
-    fn name(&self) -> &str;
-    fn is_required(&self) -> bool;
-}
-
-// ###################################################################
-// Define the RequiredArgument struct implementing the Argument trait.
-// ###################################################################
-
-pub struct RequiredArgument<T: Send + Sync> {
-    name: String,
-    _phantom: std::marker::PhantomData<T>,
-}
-
-impl<T: Send + Sync> RequiredArgument<T> {
-    pub fn new(name: &str) -> Self {
-        RequiredArgument {
-            name: name.to_string(),
-            _phantom: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<T: Send + Sync> Argument for RequiredArgument<T> {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn is_required(&self) -> bool {
-        true
-    }
-}
-
-// ###################################################################
-// Define the OptionalArgument struct implementing the Argument trait.
-// ###################################################################
-
-pub struct OptionalArgument<T> {
-    name: String,
-    _phantom: std::marker::PhantomData<T>,
-}
-
-impl<T> OptionalArgument<T> {
-    pub fn new(name: &str) -> Self {
-        OptionalArgument {
-            name: name.to_string(),
-            _phantom: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<T: Send + Sync> Argument for OptionalArgument<T> {
-    fn name(&self) -> &str {
-        &self.name
-    }
-
-    fn is_required(&self) -> bool {
-        false
-    }
-}
+use std::io::{Error, ErrorKind};
 
 // ###################################################################
 // Structure for a final parsed command.
@@ -97,27 +14,13 @@ pub struct Command<'a> {
 }
 
 impl<'a> Command<'a> {
-    const INVALID_CMD: &str = "You have entered an invalid Command!";
-    const NO_CMD: &str = "Command was not specified!";
+    const INVALID_CMD: &'static str = "You have entered an invalid Command!";
+    const NO_CMD: &'static str = "Command was not specified!";
     // const INVALID_SYNTAX: &str = "Sorry, I could not understand you.";
-    const MISSING_ARGS: &str = "Missing required variables.";
+    const MISSING_ARGS: &'static str = "Missing required variables.";
 
     pub fn invalid(error: &str) -> Error {
         Error::new(ErrorKind::InvalidData, format!("{}", error))
-    }
-
-    pub async fn callback<F, Fut>(
-        &self,
-        client: WebexClient,
-        message: Message,
-        required_argument: ArgTuple,
-        optional_arguments: ArgTuple,
-        f: F,
-    ) where
-        F: Fn(WebexClient, Message, ArgTuple, ArgTuple) -> Fut,
-        Fut: Future<Output = ()> + Send + 'static,
-    {
-        tokio::spawn(f(client, message, required_argument, optional_arguments));
     }
 }
 
@@ -157,9 +60,9 @@ impl<'a> Parser {
     // Retrieve the set of arguments required for proper command execution.
     // ------------------------------------------------------------------------------
 
-    pub fn get_command_arguments(&self, name: &str) -> &Vec<Box<dyn Argument>> {
-        &self.commands.get(name).unwrap().1
-    }
+    // pub fn get_command_arguments(&self, name: &str) -> &Vec<Box<dyn Argument>> {
+    //     &self.commands.get(name).unwrap().1
+    // }
 
     // ------------------------------------------------------------------------------
     // Parse the plain text string values into a usable command.
@@ -171,14 +74,14 @@ impl<'a> Parser {
         let num_parts = parts.len();
         if num_parts <= 1 {
             error!("No command has been specified!");
-            Err::<Command, Error>(Command::invalid(Command::NO_CMD));
+            return Err::<Command, Error>(Command::invalid(Command::NO_CMD));
         }
 
         // If the command is correctly initialized, check if it is available as
         // a key within the hasmap.
         if !self.commands.contains_key(parts[1]) {
             error!("Command not found!");
-            Err::<Command, Error>(Command::invalid(Command::INVALID_CMD));
+            return Err::<Command, Error>(Command::invalid(Command::INVALID_CMD));
         }
 
         // If the commands is present within the registered commands, retrive the
@@ -208,7 +111,7 @@ impl<'a> Parser {
             }
         } else {
             error!("Did not specified all the required arguments to execute this command!");
-            Err::<Command, Error>(Command::invalid(Command::MISSING_ARGS));
+            return Err::<Command, Error>(Command::invalid(Command::MISSING_ARGS));
         }
 
         debug!("Calling the callback function supplied on the argument list.");
